feat: check permission for update and delete projects

+++ src/controller/ProjectsController.js
@@ -75,6 +75,14 @@ class ProjectController {
             }
 
             await run();
+
+            //Check if the user trying to update the service has permission            
+            const resultPermissionDelete = await collection.findOne({ "_id": req.params.id }, { projection: { "_id": 0, "id_user": 1 } });
+
+            if (resultPermissionDelete != req.body.id_user) {
+                return res.status(400).json({ message: "You do not have permission to delete this data!" })
+            }
+
             const result = await collection.updateOne({ "_id": id_Project }, { $set: modifiedProject });
             await closeBd();
 
@@ -93,6 +101,14 @@ class ProjectController {
         try {
             await run();
             const Project = new ObjectId(req.params.id);
+
+            //Check if the user trying to update the service has permission            
+            const resultPermissionDelete = await collection.findOne({ "_id": req.params.id }, { projection: { "_id": 0, "id_user": 1 } });
+
+            if (resultPermissionDelete != req.body.id_user) {
+                return res.status(400).json({ message: "You do not have permission to delete this data!" })
+            }
+
             const result = await collection.deleteOne({ "_id": Project });
             await closeBd();
             if (result.deletedCount == 0) {
